# test-results/2025-09-03/crawler-tests/data/openapi_schema_170814.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "历史文本项目 - 数据源服务",
    "description": "\n    ## 数据源服务 API\n    \n    提供多平台内容爬取、代理管理和内容处理功能的微服务。\n    \n    ### 主要功能\n    - 🕷️ **爬虫管理**: 支持今日头条、百家号、小红书等平台\n    - 📄 **内容管理**: 手动添加、批量导入、查询和更新内容\n    - 🌐 **代理管理**: 代理池管理、自动测试和轮换\n    - 📊 **监控统计**: 实时状态监控和数据统计\n    \n    ### 技术栈\n    - **框架**: FastAPI + Python 3.9+\n    - **数据库**: MongoDB + Redis\n    - **爬虫**: Scrapy + Selenium\n    - **代理**: 多供应商代理池\n    \n    ### 使用说明\n    1. 使用 `/crawlers/` 接口管理爬虫任务\n    2. 使用 `/content/` 接口管理内容数据\n    3. 使用 `/proxy/` 接口管理代理设置\n    4. 查看 `/health` 检查服务状态\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/proxy/": {
      "get": {
        "tags": [
          "代理管理"
        ],
        "summary": "获取代理列表",
        "description": "获取代理列表，支持过滤和分页",
        "operationId": "get_proxy_list_api_v1_proxy__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProxyStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态过滤",
              "title": "Status"
            },
            "description": "按状态过滤"
          },
          {
            "name": "quality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProxyQuality"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按质量过滤",
              "title": "Quality"
            },
            "description": "按质量过滤"
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按供应商过滤",
              "title": "Provider"
            },
            "description": "按供应商过滤"
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按国家过滤",
              "title": "Country"
            },
            "description": "按国家过滤"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Size"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Proxy List Api V1 Proxy  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proxy/active": {
      "get": {
        "tags": [
          "代理管理"
        ],
        "summary": "获取可用代理",
        "description": "获取当前可用的代理列表",
        "operationId": "get_active_proxies_api_v1_proxy_active_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "返回数量限制",
              "default": 10,
              "title": "Limit"
            },
            "description": "返回数量限制"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Active Proxies Api V1 Proxy Active Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/proxy/test": {
      "post": {
        "tags": [
          "代理管理"
        ],
        "summary": "测试代理",
        "description": "测试指定代理的可用性",
        "operationId": "test_proxy_api_v1_proxy_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxyTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Test Proxy Api V1 Proxy Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/proxy/{proxy_id}/ban": {
      "post": {
        "tags": [
          "代理管理"
        ],
        "summary": "禁用代理",
        "description": "手动禁用指定的代理",
        "operationId": "ban_proxy_api_v1_proxy__proxy_id__ban_post",
        "parameters": [
          {
            "name": "proxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proxy Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "禁用原因",
              "title": "Reason"
            },
            "description": "禁用原因"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Ban Proxy Api V1 Proxy  Proxy Id  Ban Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proxy/{proxy_id}/unban": {
      "post": {
        "tags": [
          "代理管理"
        ],
        "summary": "解禁代理",
        "description": "解禁被禁用的代理",
        "operationId": "unban_proxy_api_v1_proxy__proxy_id__unban_post",
        "parameters": [
          {
            "name": "proxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proxy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Unban Proxy Api V1 Proxy  Proxy Id  Unban Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proxy/{proxy_id}": {
      "delete": {
        "tags": [
          "代理管理"
        ],
        "summary": "删除代理",
        "description": "从系统中删除指定的代理",
        "operationId": "delete_proxy_api_v1_proxy__proxy_id__delete",
        "parameters": [
          {
            "name": "proxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proxy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Proxy Api V1 Proxy  Proxy Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProxyQuality": {
        "type": "string",
        "enum": [
          "unknown",
          "low",
          "medium",
          "high",
          "premium"
        ],
        "title": "ProxyQuality",
        "description": "代理质量等级"
      },
      "ProxyStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "testing",
          "banned",
          "timeout",
          "error"
        ],
        "title": "ProxyStatus",
        "description": "代理状态枚举"
      },
      "ProxyTestRequest": {
        "properties": {
          "host": {
            "type": "string",
            "title": "Host",
            "description": "代理主机地址"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port",
            "description": "代理端口"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "用户名"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "密码"
          },
          "protocol": {
            "type": "string",
            "pattern": "^(http|https|socks4|socks5)$",
            "title": "Protocol",
            "description": "协议类型",
            "default": "http"
          }
        },
        "type": "object",
        "required": [
          "host",
          "port"
        ],
        "title": "ProxyTestRequest",
        "description": "代理测试请求模型"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}